attempt to decompile osCreateScheduler

+++ src/sched.h
@@ -83,7 +83,7 @@ void __scMain(OSSched *arg);
 void __scYield(OSSched *sc) ;
 void __scAppendList(OSSched *sc, OSScTask *t);
 void __scExec(OSSched *sc, OSScTask *sp, OSScTask *dp);
-void osCreateScheduler(OSSched *s, void *stack, u8 mode, u8 numFields);
+void osCreateScheduler(OSSched *s, void *stack, u8 mode, u32 numFields);
 void osScAddClient(OSSched *s, OSScClient *c, OSMesgQueue *msgQ, OSScClient *next);
 void osScRemoveClient(OSSched *s, OSScClient *c);
 OSMesgQueue *osScGetCmdQ(OSSched *s);
